SRAM test structure from FlowSpace project
Test structure that should allow to measure metric on single SRAM
cells. It contains the design together with documentation on
SRAM design rules and DRC violations to be waived. The documentation
linked from the README file should also contain all information in
order to be able to do the measurements on the test structures.